requirements.txt
Google_Images_Search==1.3.10
numpy==1.19.2
numpy==1.20.1
opencv_contrib_python==4.5.3.56
pandas==1.1.5
pandas==1.2.4
Pillow==8.4.0
Pillow==8.0.0
Pillow==8.3.1
Pillow==8.2.0
Pillow==10.0.1
pyTelegramBotAPI==2.2.3
pyTelegramBotAPI==4.14.0
python-telegram-bot==20.5
